+++ src/main/java/com/danrus/pas/PlayerArmorStandsClient.java
@@ -5,6 +5,7 @@ import com.danrus.pas.api.reg.InfoTranslators;
 import com.danrus.pas.impl.features.*;
 import com.danrus.pas.impl.holder.CapeData;
 import com.danrus.pas.impl.holder.SkinData;
+import com.danrus.pas.impl.translators.cape.MinecraftCapesTranslator;
 import com.danrus.pas.impl.translators.cape.MojangCapeTranslator;
 import com.danrus.pas.impl.translators.cape.NamemcCapeTranslator;
 import com.danrus.pas.impl.translators.skin.FileSkinTranslator;
@@ -22,12 +23,12 @@ public class PlayerArmorStandsClient {
 
     public static void initialize() {
         PasManager.getInstance();
-        FeatureRegistry fratures = FeatureRegistry.getInstance();
-        fratures.register(OverlayFeature.class);
-        fratures.register(CapeFeature.class);
-        fratures.register(SkinProviderFeature.class);
-        fratures.register(SlimFeature.class);
-        fratures.register(DisplayNameFeature.class);
+        FeatureRegistry features = FeatureRegistry.getInstance();
+        features.register(OverlayFeature.class);
+        features.register(CapeFeature.class);
+        features.register(SkinProviderFeature.class);
+        features.register(SlimFeature.class);
+        features.register(DisplayNameFeature.class);
 
         InfoTranslators translator = InfoTranslators.getInstance();
         translator.register(SkinData.class, new NamemcSkinTranslator());
@@ -35,5 +36,6 @@ public class PlayerArmorStandsClient {
         translator.register(SkinData.class, new FileSkinTranslator());
         translator.register(CapeData.class, new NamemcCapeTranslator());
         translator.register(CapeData.class, new MojangCapeTranslator());
+        translator.register(CapeData.class, new MinecraftCapesTranslator());
     }
 }