+++ src/main/java/com/danrus/pas/managers/CapeProvidersManager.java
@@ -5,13 +5,14 @@ import com.danrus.pas.api.info.NameInfo;
 import com.danrus.pas.impl.features.CapeFeature;
 import com.danrus.pas.impl.holder.CapeData;
 import com.danrus.pas.impl.providers.MojangProvider;
+import com.danrus.pas.impl.providers.cape.MinecraftCapesTextureProvider;
 import com.danrus.pas.impl.providers.cape.NamemcCapeProvider;
 import com.danrus.pas.impl.providers.common.AbstractTextureProviderManager;
 
 public class CapeProvidersManager extends AbstractTextureProviderManager<CapeData> {
 
     private static final String DEFAULT_LITERAL = "M";
-    private static final String EXCLUDE_LITERALS = "F";
+    private static final String EXCLUDE_LITERALS = "FI";
 
     @Override
     protected String getOutputString(NameInfo info) {
@@ -21,6 +22,7 @@ public class CapeProvidersManager extends AbstractTextureProviderManager<CapeDat
     @Override
     protected void prepareProviders() {
         this.addProvider(MojangProvider.getInstance());
+        this.addProvider(new MinecraftCapesTextureProvider(), 2);
         this.addProvider(new NamemcCapeProvider(), 1);
     }
 
@@ -47,5 +49,10 @@ public class CapeProvidersManager extends AbstractTextureProviderManager<CapeDat
     @Override
     protected DataRepository<CapeData> getDataManager() {
         return PasManager.getInstance().getCapeDataManager();
+    }
+
+    @Override
+    protected boolean registerAsExistingProvider() {
+        return false;
     }
 }