+++ src/main/java/com/danrus/pas/render/armorstand/ModelWithCape.java
@@ -1,7 +1,0 @@
-package com.danrus.pas.render.armorstand;
-
-import net.minecraft.client.model.geom.ModelPart;
-
-public interface ModelWithCape {
-    ModelPart getCape();
-}