+++ src/main/java/com/danrus/pas/render/armorstand/PlayerArmorStandModel.java
@@ -7,18 +7,24 @@ import com.danrus.pas.config.ModConfig;
 import com.danrus.pas.impl.holder.SkinData;
 import com.danrus.pas.managers.PasManager;
 import com.danrus.pas.utils.VersioningUtils;
+import com.mojang.blaze3d.vertex.PoseStack;
+import com.mojang.blaze3d.vertex.VertexConsumer;
 import net.minecraft.client.model.ArmorStandArmorModel;
 import net.minecraft.client.model.HumanoidModel;
 import net.minecraft.client.model.geom.ModelPart;
 import net.minecraft.client.model.geom.PartPose;
 import net.minecraft.client.model.geom.builders.*;
+import net.minecraft.client.renderer.MultiBufferSource;
+import net.minecraft.client.renderer.RenderType;
+import net.minecraft.client.renderer.texture.OverlayTexture;
 import net.minecraft.core.Rotations;
 import net.minecraft.network.chat.Component;
+import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.entity.decoration.ArmorStand;
 
 import java.util.List;
 
-public class PlayerArmorStandModel extends ArmorStandArmorModel implements ModelWithCape {
+public class PlayerArmorStandModel extends ArmorStandArmorModel implements Cape {
     // WIDE
     public final ModelPart leftSleeve;
     public final ModelPart rightSleeve;
@@ -213,7 +219,7 @@ public class PlayerArmorStandModel extends ArmorStandArmorModel implements Model
         this.leftEar.visible = isEarsVisible;
         this.rightEar.visible = isEarsVisible;
 
-        this.setModelVisibility(!showArmorStandWhileDownload(customName, PasManager.getInstance().findSkinData(info)), info.wantBeSlim(), showBase);
+        this.setModelVisibility(!showArmorStandWhileDownload(PasManager.getInstance().findSkinData(info)), info.wantBeSlim(), showBase);
 
         if (customNameString.isEmpty() && ModConfig.get().defaultSkin.isEmpty()) {
             setOriginalAngles(showBase, showArms, bodyPose);
@@ -287,13 +293,12 @@ public class PlayerArmorStandModel extends ArmorStandArmorModel implements Model
         this.shoulderStick.zRot = ((float)Math.PI / 180F) * VersioningUtils.getZRot(bodyPose);
     }
 
-    @Override
     public ModelPart getCape() {
         return this.cloak;
     }
 
 
-    public static boolean showArmorStandWhileDownload(Component customName, DataHolder data) {
+    public static boolean showArmorStandWhileDownload(DataHolder data) {
 
         if (data == null) {
             return true;
@@ -306,5 +311,14 @@ public class PlayerArmorStandModel extends ArmorStandArmorModel implements Model
 
     private static void cpp(ModelPart from, ModelPart to) {
         VersioningUtils.copyPartPose(from, to);
+    }
+
+    @Override
+    public void draw(PoseStack stack, ResourceLocation textureLocation, RenderVersionContext context, int i) {
+        //? <1.21.9 {
+        getCape().visible = true;
+        VertexConsumer vertexConsumer = context.getData(MultiBufferSource.BufferSource.class).getBuffer(RenderType.entitySolid(textureLocation));
+        getCape().render(stack, vertexConsumer, i, OverlayTexture.NO_OVERLAY);
+        //?}
     }
 }